chatservice/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
chromadb==0.4.22
langchain==0.1.3
langchain_openai==0.0.3
openai==1.9.0
pypdf==4.0.0
python-dotenv==1.0.1
redis==5.0.1
tiktoken==0.5.2

